+++ products/dgps/icd/include/BestPosMessage.h
@@ -21,7 +21,7 @@ protected:
 	E_MESSAGE_ID_INPUT_DGPS_DLV3 GetMessageID() const override;
 
 public:
-	BestPosMessage(int hertz);
+	BestPosMessage(float hertz);
 
 	virtual ~BestPosMessage() = default;
