+++ products/idan/icd/src/IdanMessage.cpp
@@ -8,5 +8,5 @@
 #include "IdanMessage.h"
 #include "LoggerProxy.h"
 
-IdanMessage::IdanMessage(int hertz): IMessage(hertz) {
+IdanMessage::IdanMessage(float hertz): IMessage(hertz) {
 }
